backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
sqlalchemy==2.0.25
pydantic==2.5.3
pydantic[email]==2.5.3
python-multipart==0.0.6
pyjwt==2.8.0
